Move `PostSignScreen` out of `OfferViewModel`
