+++ app/src/main/java/com/hedvig/app/feature/offer/OfferViewModel.kt
@@ -4,7 +4,6 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.hedvig.android.owldroid.graphql.OfferQuery
 import com.hedvig.android.owldroid.graphql.RedeemReferralCodeMutation
-import com.hedvig.android.owldroid.type.QuoteBundleAppConfigurationPostSignStep
 import com.hedvig.android.owldroid.type.QuoteBundleAppConfigurationTitle
 import com.hedvig.android.owldroid.type.SignMethod
 import com.hedvig.app.authenticate.LoginStatus
@@ -65,20 +64,6 @@ abstract class OfferViewModel : ViewModel() {
         object DiscardOffer : Event()
     }
 
-    enum class PostSignScreen {
-        CONNECT_PAYIN,
-        MOVE,
-        CROSS_SELL;
-
-        companion object {
-            fun from(postSignStep: QuoteBundleAppConfigurationPostSignStep) = when (postSignStep) {
-                QuoteBundleAppConfigurationPostSignStep.CONNECT_PAYIN -> CONNECT_PAYIN
-                QuoteBundleAppConfigurationPostSignStep.MOVE -> MOVE
-                QuoteBundleAppConfigurationPostSignStep.CROSS_SELL -> CROSS_SELL
-                QuoteBundleAppConfigurationPostSignStep.UNKNOWN__ -> CONNECT_PAYIN
-            }
-        }
-    }
 
     protected val _events = MutableSharedFlow<Event>(
         extraBufferCapacity = 1,
@@ -310,3 +295,4 @@ class OfferViewModelImpl(
         }
     }
 }
+
